+++ crates/amaru-ledger/src/state/volatile/db.rs
@@ -1262,6 +1262,7 @@ mod tests {
             protocol_parameters: PREPROD_DEFAULT_PROTOCOL_PARAMETERS.clone(),
             pruned_proposals: BTreeMap::new(),
             payouts: BTreeMap::new(),
+            treasury_withdrawals: 0,
             is_dormant_epoch: false,
             constitutional_committee: committee,
             new_constitution: None,